+++ apps/backoffice-v2/src/domains/individuals/queries/index.ts
@@ -1,0 +1,2 @@
+export * from './useEndUserByIdQuery';
+export * from './useEndUsersByIdsQuery';