+++ sdks/web-ui-sdk/CHANGELOG.md
@@ -1,5 +1,12 @@
 # web-ui-sdk
 
+## 1.5.92
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.91
+
 ## 1.5.91
 
 ### Patch Changes